refactored following refactor from upstream
